+++ src/main/kotlin/xyz/jpenilla/toothpick/shadow/ModifiedLog4j2PluginsCacheFileTransformer.kt
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package xyz.jpenilla.toothpick.transformer
+package xyz.jpenilla.toothpick.shadow
 
 import com.github.jengelman.gradle.plugins.shadow.transformers.Log4j2PluginsCacheFileTransformer
 import com.github.jengelman.gradle.plugins.shadow.transformers.Transformer
